+++ decoders/token-2022-decoder/src/instructions/graphql/set_transfer_fee_schema.rs
@@ -25,7 +25,7 @@ impl TryFrom<crate::instructions::postgres::SetTransferFeeRow> for SetTransferFe
             transfer_fee_discriminator: carbon_core::graphql::primitives::U8(
                 (*row.transfer_fee_discriminator) as u8,
             ),
-            transfer_fee_basis_points: (*row.transfer_fee_basis_points) as i32,
+            transfer_fee_basis_points: *row.transfer_fee_basis_points,
             maximum_fee: carbon_core::graphql::primitives::U64(*row.maximum_fee),
         })
     }
